+++ src/index.js
@@ -1,11 +1,3 @@
-import TerraformDrawer from 'src/draw/TerraformDrawer';
-import TerraformMetadata from 'src/metadata/TerraformMetadata';
-import TerraformParser from 'src/parser/TerraformParser';
-import TerraformRenderer from 'src/render/TerraformRenderer';
+import TerraformPlugin from 'src/models/TerraformPlugin';
 
-export default {
-  PluginDrawer: TerraformDrawer,
-  PluginMetadata: TerraformMetadata,
-  PluginParser: TerraformParser,
-  PluginRenderer: TerraformRenderer,
-};
+export default TerraformPlugin;