+++ src/render/TerraformRenderer.js
@@ -2,7 +2,6 @@ import nunjucks from 'nunjucks';
 import template from 'src/render/TerraformTemplate';
 import {
   DefaultRender,
-  ComponentAttribute,
   FileInput,
 } from 'leto-modelizer-plugin-core';
 
@@ -13,8 +12,8 @@ class TerraformRenderer extends DefaultRender {
   /**
    * Default constructor, initialize nunjucks library and template.
    */
-  constructor() {
-    super();
+  constructor(pluginData) {
+    super(pluginData);
     this.template = template;
     nunjucks.configure({
       autoescape: false,
@@ -25,65 +24,13 @@ class TerraformRenderer extends DefaultRender {
 
   /**
    * Convert all provided components and links in terraform files.
-   * @param {Component[]} componentsTree - List of components you want to convert.
-   * @param {ComponentLink[]} links - List of links you want to convert.
-   * @param {String} defaultFileName - Default file name to set in case of empty path on component.
    * @return {FileInput[]} - Array of generated files from components and links.
    */
-  render(componentsTree, links, defaultFileName = './new_file.tf') {
-    links.forEach((link) => {
-      const component = componentsTree.find((c) => c.id === link.source);
-      if (!component) {
-        return;
-      }
-      this.setComponentAttribute(
-        component,
-        link.definition.attributeRef,
-        'Array',
-        link.target,
-      );
-    });
-
+  render() {
     const componentsMap = new Map();
-    this.collectComponentsFromTree(componentsMap, componentsTree, defaultFileName);
+    this.collectComponentsFromTree(componentsMap, this.pluginData.components, './new_file.tf');
 
     return this.generateFilesFromComponentsMap(componentsMap);
-  }
-
-  /**
-   * Set attribute value on component.
-   * @param {Component} component - Component to set attribute.
-   * @param {String} name - Name of attribute to set.
-   * @param {String} type - Type of attribute.
-   * @param {String} value - Value of attribute, only id of component is expected here.
-   */
-  setComponentAttribute(component, name, type, value) {
-    const attribute = component.attributes.find((a) => a.name === name);
-    const isArray = type === 'Array';
-
-    if (!attribute) {
-      component.attributes.push(new ComponentAttribute({
-        name,
-        value: isArray ? [value] : value,
-        type,
-      }));
-    } else if (isArray && !attribute.value.includes(value)) {
-      attribute.value.push(value);
-    } else if (!isArray) {
-      attribute.value = value;
-    }
-  }
-
-  /**
-   * Set container parent id on child attribute.
-   * @param {Component} parent - Parent component to get id.
-   * @param {Component} child - Child to set id attribute with parent id.
-   */
-  setContainerAttribute(parent, child) {
-    const definition = child.definition.definedAttributes
-      .find((def) => def.type === 'Reference' && def.containerRef === parent.definition.type);
-
-    this.setComponentAttribute(child, definition.name, 'String', parent.id);
   }
 
   /**
@@ -105,7 +52,7 @@ class TerraformRenderer extends DefaultRender {
       if (component.children.length === 0) {
         return;
       }
-      component.children.forEach((child) => this.setContainerAttribute(component, child));
+
       this.collectComponentsFromTree(files, component.children, component.path);
     });
   }