+++ tests/resources/tf/link_reverse_single.tf
@@ -1,6 +1,6 @@
 resource "parent" "parent_reverse_single_1" {
     name = "parent_reverse_single_1"
-    fromChild = "child_reverse_single_1"
+    fromChild = ["child_reverse_single_1"]
 }
 
 resource "child" "child_reverse_single_1" {