+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+
+### Fixed
+
+- Add missing icon for aws provider and aws db instance
+- Fix oversize of all icons
+- Fix container metadata in aws metadata
+- Fix empty display for aws provider
+- Add default values to metadata properties `attributes` and `isContainer`
+
 ## [0.1.2] - 2022/10/05
 
 ### Fixed