+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [Unreleased]
+## [0.1.6]
 
 ### Changed
 
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add renderer for terraform files
 - Import metadata instead of passing it through constructor in TerraformMetadata
 
+[0.1.6]: https://github.com/ditrit/terrator-plugin/blob/0.1.6/changelog.md
 [0.1.5]: https://github.com/ditrit/terrator-plugin/blob/0.1.5/changelog.md
 [0.1.4]: https://github.com/ditrit/terrator-plugin/blob/0.1.4/changelog.md
 [0.1.3]: https://github.com/ditrit/terrator-plugin/blob/0.1.3/changelog.md